requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-filter==2.4.0
gunicorn==20.0.4
idna==2.10
jsonfield==3.1.0
numpy==1.20.1
pandas==1.2.2
pbr==5.5.1
pybliometrics==2.9.0
pyodbc==4.0.30
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
